requirements.txt
# Dependensi Python
checkov
pyyaml
pytest